tests addInfo, deletePost, isValidField, signIn
